requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
Bottleneck
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-ckeditor==6.7.2
django-js-asset==3.0.1
flatbuffers==25.2.10
fqdn==1.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
idna==3.10
ipykernel==6.29.5
isoduration==20.11.0
joblib
jupyter==1.1.1
jupyterlab==4.3.5
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft
mkl_random
ml-dtypes==0.4.1
namex==0.0.8
notebook==7.3.2
numexpr
numpy
opt_einsum==3.4.0
optree==0.14.0
packaging
pandas
patsy==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
plotly
protobuf==5.29.3
Pygments==2.19.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil
pytz
requests==2.32.3
rich==13.9.4
scikit-learn
scipy
setuptools==75.8.0
six
sqlparse==0.5.3
statsmodels==0.14.4
tenacity
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl
typing_extensions==4.12.2
tzdata
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.3.0
webcolors==24.11.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
